+++ packages/mythic-configuration/src/index.ts
@@ -4,8 +4,3 @@ export { setConfigFile } from "./backends/filesystem"
 export { configuration } from "./package";
 export { createConfigOption } from "./create-option";
 export * from "./types";
-
-export const {
-  selector: theme,
-  action: setTheme,
-} = createConfigOption("theme")("light");
