+++ packages/notebook-app-component/src/notebook-apps/draggable-with-cell-creator.tsx
@@ -1,26 +1,17 @@
 import { CellType } from "@nteract/commutable";
 import { ContentRef } from "@nteract/core";
-import {
-  Cells,
-  CodeCell,
-  MarkdownCell,
-  RawCell
-} from "@nteract/stateful-components";
+import { Cells, CodeCell, MarkdownCell, RawCell } from "@nteract/stateful-components";
+import { ThemableNotifications, Themer } from "@nteract/stateful-components";
 import React from "react";
-
-import StatusBar from "../derived-components/status-bar";
-import { ThemableNotifications } from "../derived-components/themable-notifications";
-import CellToolbar from "../derived-components/toolbar";
-
 import { DragDropContext as dragDropContext } from "react-dnd";
 import HTML5Backend from "react-dnd-html5-backend";
-
 import CellCreator from "../decorators/cell-creator";
 import DraggableCell from "../decorators/draggable";
 import HijackScroll from "../decorators/hijack-scroll";
 import KeyboardShortcuts from "../decorators/kbd-shortcuts";
-import Themer from "../decorators/themer";
 import UndoableCellDelete from "../decorators/undoable/undoable-cell-delete";
+import StatusBar from "../derived-components/status-bar";
+import CellToolbar from "../derived-components/toolbar";
 
 interface ComponentProps {
   contentRef: ContentRef;
