+++ packages/stateful-components/src/notebook/themable-notifications.tsx
@@ -1,9 +1,8 @@
-import { selectors } from "@nteract/core";
-import { theme } from "@nteract/mythic-configuration";
 import { NotificationRoot } from "@nteract/mythic-notifications";
 import { AppState } from "@nteract/types";
 import React from "react";
 import { connect } from "react-redux";
+import { getTheme } from "../decorators/themer";
 
 interface Props {
   theme: string;
@@ -13,6 +12,6 @@ const PureThemableNotifications = ({ theme }: Props) =>
   <NotificationRoot darkTheme={theme === "dark"}/>;
 
 export const ThemableNotifications = connect(
-  (state: AppState) => ({theme: theme(state)}),
+  (state: AppState) => ({theme: getTheme(state)}),
 )(PureThemableNotifications);
 ThemableNotifications.displayName = "ThemableNotifications";
